use trait hydrator with hydrate method
